requirements.txt
# Core AI / Voice
google-genai>=0.8.0
google-cloud-storage>=2.16.0

# PDF Processing
pymupdf>=1.24.0

# Frontend
streamlit>=1.35.0
plotly>=5.22.0

# Utilities
python-dotenv>=1.0.0
pyaudio>=0.2.11